Added Area Cell Setup
Added area cell setup for the game world. I'm going to change the camera system to use this instead so that I don't have to set up triggers. I'm not sure if I want to have a camera that doesn't just focus on one area, but either way I likely won't use it, so eh.