Map Input Change + Block Fix + Working on Text Box
I changed the map input so that the map itself checks for the 'close button' instead of the player script. I did this in case I want to have the player be able to access buttons using the keyboard.

I fixed the block glitch where all blocks didn't check for the type of weapon being used on them.

I'm also working on having text boxes to explain elements in the game. I should also have a toggle for turning the notifications off.