Redid Game Camera and Optimized Game World
I redid the game camera so that it can move dynamically, but right now the camera is still fixed. I'm not sure if I'll have the game camera view get changed or not.

I also optimized the game world so that unneeded areas are deactivated. However, this causes a delay when the player loads in from a portal, or when a player dies. Basically, any event that causes the player to go to an area not adjacent to their current area causes this stutter. I intend to have a death animation, so I think I may be able to mask that, but it's something to keep track of.