Working on the Loading Scene for the Game
Working on implementing a loading scene for the game. It seems to still stutter when the title scene enters the loading scene, and the load just jumps to 100%. I need to figure out why this happens.

Also, I'll probably just have a general loading animation instead of showing the percentage being loaded.