Added Properly Looping BGMs
I added the crossfader so that the BGMs loop properly. I haven't tested it fully, but it should be fine. Also, the async scene loader appears to be fine? I hope the way I set up the loading screen doesn't cause issues, but I may have to look into it more.