Level File Change + Bullet Changes + Time SFX Change + Tutorial Fix
I changed the B00 file because one of the scraps was spawning inside a wall. I also made some changes to the stats of the bullets in the game. I need to make sure that B00 is updated on all my systems, and that the world is rebuilt.

On another note, I had the ticking sound play at 5 seconds instead of at 10 seconds. I also fixed some typos in the tutorial text.