Built the Game + Some Slight Resolution Fixes
Built the game for WebGL and Windows. I also fixed the resolution issue by changing the canvas settings in the loading scene (it was not the same as everything else).

There is still a glitch where the camera doesn't adjust right away, and where the game can sometimes lock the player into a tutorial when starting a new game, even though the tutorials are turned off (opening settings fixes this issue). I think this has to do with hold-overs from prior loaded versions of the game scene, which if so I can't really fix right now. I think I'll leave the game as is and just make a note of the problems.